+++ server/src/models/errors/error-type.ts
