+++ server/src/models/results/comparePwResult.ts
@@ -1,6 +1,5 @@
 import { SocoboUser } from "./../database/socobouser";
 
-
 export class ComparePwResult {
   constructor (public isPasswordMatch: boolean, public user: SocoboUser) {}
 }
