+++ server/src/models/results/loginResult.ts
@@ -1,6 +1,5 @@
 import { SocoboUser } from "./../database/socoboUser";
 
-
 export class LoginResult {
   private token: string;
   private user: SocoboUser;
